Implement rosefire login with authservice, add profile setup page
The login page now has rosefire implemented and working with an
authservice that can be accessed from any page in the application.

The login and profileSetup pages have a little bit of styling/html, but
they will certainly need more cleaning up later.

This also adds an image picker native plugin which will only work if
you run the app on a simulator (untested).

This also adds the angularfire2 module for later use.