Only register on first login
This changes the keys stored in FB to be the UID of the authState
object rather than auto-generated. This way, we can check if the user
has already created an account on login and push directly to the
TabPage.

This also enhances the phone # input field to be much more user
freindly (and make it more consistent for us 👍)